End of section 10

+++ Application/Activities/DeleteActivity.cs
@@ -1,6 +1,8 @@
 using System;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Errors;
 using MediatR;
 using Persistence;
 
@@ -30,6 +32,11 @@ namespace Application.Activities
             {
                 // create object here to save
                 var activity = await context.Activities.FindAsync(request.Id);
+
+                if(activity == null) {
+                    throw new RestException(HttpStatusCode.NotFound, new {activity = "Not Found"});
+                }
+
                 context.Activities.Remove(activity);
                 var success = await context.SaveChangesAsync() > 0;
